requirements.txt
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
langgraph 
langchain-openai
groq
nest-asyncio
requests
python-dotenv
google-auth